# server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the library store server for manage books, authors and reservations"
  version: "1.0.0"
  title: "Library store"
  termsOfService: "http://abey-calabro.heroku.com"
  contact:
    email: "apiteam@abey-calabro.heroku.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "abey-calabro.heroku.com"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /themes:
    get:
      tags:
      - "theme"
      summary: "Get all themes"
      description: "Return all themes in the DB"
      operationId: "getThemes"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Start offset"
        required: false
        type: "integer"
        format: "int32"
      - name: "lenght"
        in: "query"
        description: "Result lenght"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successfull operation"
          schema:
            $ref: "#/definitions/Theme"
      x-swagger-router-controller: "Theme"
  /theme/findByName:
    get:
      tags:
      - "theme"
      summary: "Get genres by theme"
      description: "Return all themes by a specific name"
      operationId: "getThemesByName"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "The name to filter by"
        required: true
        type: "string"
      - name: "offset"
        in: "query"
        description: "Start offset"
        required: false
        type: "integer"
        format: "int32"
      - name: "lenght"
        in: "query"
        description: "Result lenght"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successfull operation"
          schema:
            $ref: "#/definitions/Theme"
        400:
          description: "Invalid name value"
      x-swagger-router-controller: "Theme"
...
definitions:
...
  Theme:
    type: "object"
    properties:
      name:
        type: "string"
    example:
      name: "name"
...